requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
numpy
pandas
scikit-learn
matplotlib